+++ R/colour_values_rgb.R
@@ -116,7 +116,7 @@ colour_values_to_rgb.character <- function( x, palette, na_colour, alpha, includ
 #' }
 #'
 #' @export
-colour_values_to_rgb.default <- function( x, palette, na_colour, alpha, include_alpha, n_summaries = 0, format = FALSE, digits = 2 ) {
+colour_values_to_rgb.default <- function( x, palette, na_colour, alpha, include_alpha, n_summaries = 0, format = TRUE, digits = 2 ) {
   colour_num_values_with_palette_rgb( palette, x, na_colour, alpha, include_alpha, n_summaries, format, "numeric", digits )
 }
 
@@ -140,19 +140,19 @@ colour_values_to_rgb.factor <-  function( x, palette, na_colour, alpha, include_
 
 #' @rdname colour_values_rgb
 #' @export
-colour_values_to_rgb.Date <-  function( x, palette, na_colour, alpha, include_alpha, n_summaries = 0, format  = FALSE ) {
+colour_values_to_rgb.Date <-  function( x, palette, na_colour, alpha, include_alpha, n_summaries = 0, format  = TRUE ) {
   colour_num_values_with_palette_rgb( palette, x, na_colour, alpha, include_alpha, n_summaries, format, "Date", 0 )
 }
 
 #' @rdname colour_values_rgb
 #' @export
-colour_values_to_rgb.POSIXct <-  function( x, palette, na_colour, alpha, include_alpha, n_summaries = 0, format = FALSE ) {
+colour_values_to_rgb.POSIXct <-  function( x, palette, na_colour, alpha, include_alpha, n_summaries = 0, format = TRUE ) {
   colour_num_values_with_palette_rgb( palette, x, na_colour, alpha, include_alpha, n_summaries, format, "POSIXct", 0 )
 }
 
 #' @rdname colour_values_rgb
 #' @export
-colour_values_to_rgb.POSIXlt <-  function( x, palette, na_colour, alpha, include_alpha, n_summaries = 0, format = FALSE ) {
+colour_values_to_rgb.POSIXlt <-  function( x, palette, na_colour, alpha, include_alpha, n_summaries = 0, format = TRUE ) {
   colour_num_values_with_palette_rgb( palette, as.POSIXct(x), na_colour, alpha, include_alpha, n_summaries, format, "POSIXct", 0 )
 }