+++ NEWS.md
@@ -1,5 +1,6 @@
 # colourvalues 0.2.0
 
+* `format`, `summary`, `n_summaries` arguments for returning summary colours (for use in legends)
 * `convert_colours()` function exposed to R to convert between hex-rgb [issue 14](https://github.com/SymbolixAU/colourvalues/issues/14)
 * speed improvement for internal hex-to-RGB conversion
 * speed improvement for `NumericVector`s [issue 26](https://github.com/SymbolixAU/colourvalues/issues/26)